+++ tests/testthat/test_hy_annual_instant_peaks.R
@@ -6,13 +6,13 @@ test_that("hy_annual_instant_peaks accepts single and multiple province argument
             expect_identical(unique(
               hy_annual_instant_peaks(
                 station_number = stns,
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             ), stns)
             expect_identical(length(unique(
               hy_annual_instant_peaks(
                 station_number = c("08NM083", "08NE102"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             )), length(c("08NM083", "08NE102")))
           })
@@ -23,13 +23,13 @@ test_that("hy_annual_instant_peaks accepts single and multiple province argument
             expect_true(nrow(
               hy_annual_instant_peaks(
                 prov_terr_state_loc = "BC",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
             expect_true(nrow(
               hy_annual_instant_peaks(
                 prov_terr_state_loc = c("BC", "YT"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -38,18 +38,18 @@ test_that("hy_annual_instant_peaks produces an error when a province is not spec
           {
             expect_error(hy_annual_instant_peaks(
               prov_terr_state_loc = "BCD",
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
             expect_error(hy_annual_instant_peaks(
               prov_terr_state_loc = c("AB", "BCD"),
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
           })
 
 test_that("hy_annual_instant_peaks gather data when no arguments are supplied",
           {
             expect_true(nrow(hy_annual_instant_peaks(
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             )) >= 1)
           })
 
@@ -59,7 +59,7 @@ test_that("hy_annual_instant_peaks can accept both arguments for backward compat
               hy_annual_instant_peaks(
                 prov_terr_state_loc = "BC",
                 station_number = "08MF005",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -71,7 +71,7 @@ test_that("hy_annual_instant_peaks respects year inputs", {
   df <-
     hy_annual_instant_peaks(
       station_number = c("08NM083", "08NE102"),
-      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"),
+      hydat_path = hy_test_db(),
       start_year = s_year,
       end_year = e_year
     )