+++ tests/testthat/test_hy_daily_levels.R
@@ -6,13 +6,13 @@ test_that("hy_daily_levels accepts single and multiple province arguments",
             expect_identical(unique(
               hy_daily_levels(
                 station_number = stns,
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             ), stns)
             expect_identical(length(unique(
               hy_daily_levels(
                 station_number = c("08MF005", "05AA008"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             )), length(c("08NM083", "08NE102")))
           })
@@ -23,13 +23,13 @@ test_that("hy_daily_levels accepts single and multiple province arguments",
             expect_true(nrow(
               hy_daily_levels(
                 prov_terr_state_loc = "BC",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
             expect_true(nrow(
               hy_daily_levels(
                 prov_terr_state_loc = c("YT", "BC"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -38,17 +38,17 @@ test_that("hy_daily_levels produces an error when a province is not specified co
           {
             expect_error(hy_daily_levels(
               prov_terr_state_loc = "BCD",
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
             expect_error(hy_daily_levels(
               prov_terr_state_loc = c("AB", "BCD"),
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
           })
 
 ## Too much data
 # test_that("hy_daily_levels gather data when no arguments are supplied",{
-#  expect_true(nrow(hy_daily_levels(hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
+#  expect_true(nrow(hy_daily_levels(hydat_path = hy_test_db())) >= 1)
 # })
 
 test_that("hy_daily_levels can accept both arguments for backward compatability",
@@ -57,7 +57,7 @@ test_that("hy_daily_levels can accept both arguments for backward compatability"
               hy_daily_levels(
                 prov_terr_state_loc = "BC",
                 station_number = "08MF005",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -68,7 +68,7 @@ test_that("hy_daily_levels respects Date specification", {
   temp_df <- hy_daily_levels(
     station_number = "08MF005",
     prov_terr_state_loc = "BC",
-    hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"),
+    hydat_path = hy_test_db(),
     start_date = date_vector[1],
     end_date = date_vector[2]
   )
@@ -79,7 +79,7 @@ test_that("hy_daily_levels correctly parses leaps year", {
   expect_warning(
     hy_daily_levels(
       prov_terr_state_loc = "BC",
-      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"),
+      hydat_path = hy_test_db(),
       start_date = "1988-02-29",
       end_date = "1988-02-29"
     ),