+++ tests/testthat/test_hy_sed_samples.R
@@ -6,13 +6,13 @@ test_that("hy_sed_samples accepts single and multiple province arguments",
             expect_identical(unique(
               hy_sed_samples(
                 station_number = stns,
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             ), stns)
             expect_identical(length(unique(
               hy_sed_samples(
                 station_number = c("05AA008", "08MF005"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )$STATION_NUMBER
             )), length(c("08NM083", "08NE102")))
           })
@@ -23,13 +23,13 @@ test_that("hy_sed_samples accepts single and multiple province arguments",
             expect_true(nrow(
               hy_sed_samples(
                 prov_terr_state_loc = "BC",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
             expect_true(nrow(
               hy_sed_samples(
                 prov_terr_state_loc = c("BC", "AB"),
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -38,17 +38,17 @@ test_that("hy_sed_samples produces an error when a province is not specified cor
           {
             expect_error(hy_sed_samples(
               prov_terr_state_loc = "BCD",
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
             expect_error(hy_sed_samples(
               prov_terr_state_loc = c("BC", "BCD"),
-              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              hydat_path = hy_test_db()
             ))
           })
 
 ## Too much data
 # test_that("hy_sed_samples gather data when no arguments are supplied",{
-#  expect_true(nrow(hy_sed_samples(hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
+#  expect_true(nrow(hy_sed_samples(hydat_path = hy_test_db())) >= 1)
 # })
 
 test_that("hy_sed_samples can accept both arguments for backward compatability",
@@ -57,7 +57,7 @@ test_that("hy_sed_samples can accept both arguments for backward compatability",
               hy_sed_samples(
                 prov_terr_state_loc = "BC",
                 station_number = "08MF005",
-                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+                hydat_path = hy_test_db()
               )
             ) >= 1)
           })
@@ -68,7 +68,7 @@ test_that("hy_sed_samples respects Date specification", {
   expect_error(
     hy_sed_samples(
       station_number = "08MF005",
-      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"),
+      hydat_path = hy_test_db(),
       start_date = date_vector[1],
       end_date = date_vector[2]
     ),
@@ -80,7 +80,7 @@ test_that("hy_sed_samples correctly parses leaps year", {
   expect_warning(
     hy_sed_samples(
       prov_terr_state_loc = "BC",
-      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"),
+      hydat_path = hy_test_db(),
       start_date = "1976-02-29",
       end_date = "1976-02-29"
     ),