+++ tests/testthat/test_station_choice.R
@@ -1,7 +1,7 @@
 context("Make sure that station choice chooses the correct station")
 
 test_that("Outputs that same station that is inputted",{
-  hydat_path <- system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+  hydat_path <- hy_test_db()
   ## Read in database
   hydat_con <- DBI::dbConnect(RSQLite::SQLite(), hydat_path)
   stns <- c("08NM083", "08NE102")